% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sankeydf}
\alias{sankeydf}
\title{Example Data for Sankey Sankey Diagrams}
\format{A tibble data frame with 46 rows and 3 variables.}
\usage{
sankeydf
}
\description{
Data Source: https://www.highcharts.com/demo/sankey-diagram
}
\keyword{datasets}
